menambahkan middleware pada method post dan delete user

+++ routers/users.js
@@ -4,7 +4,7 @@ const auth = require('../middleware/auth');
 
 const router = express.Router()
 
-router.get('',auth, (req, res) => {
+router.get('', auth, (req, res) => {
     connection.query('SELECT * FROM tbl_user', (err, results, fields) => {
         if (err) throw err
         res.json(results)
@@ -12,20 +12,39 @@ router.get('',auth, (req, res) => {
 })
 
 router.post('/', (req, res) => {
+    connection.query('SELECT COUNT(*) as jumlah_user FROM tbl_user', (err, results, fields) => {
+        if (results.jumlah_user > 0) {
+            auth(req, res, next)
+        } else {
+            next()
+        }
+    })
+}, (req, res) => {
     const sql = 'INSERT INTO tbl_user (id,username,password) VALUES(?,?,?)'
-    const values = [null,req.body.username,req.body.password]
+    const values = [null, req.body.username, req.body.password]
     connection.query(sql, values, (err, results, fields) => {
-        if (err) throw err
-        res.json(results)
+        if (err) {
+            res.end(500)
+            return
+        }
+        res.json({ values })
     })
 })
 
 router.delete('/:id', auth, (req, res) => {
-    const sql = 'DELETE FROM tbl_deskripsi WHERE tbl_user id = ?'
     const values = [req.params.id]
-    connection.query(sql, values, (err, results, fields) => {
-        if (err) throw err
-        res.json(results)
+
+    connection.query('SELECT COUNT(*) as jumlah_user FROM tbl_user', (err, results, fields) => {
+        if (results.jumlah_user > 1) {
+            const sql = 'DELETE FROM tbl_deskripsi WHERE tbl_user id = ?'
+            connection.query(sql, values, (err, results, fields) => {
+                if (err) throw err
+                res.json(results)
+            })
+        } else {
+            console.log("Tidak Bisa Di hapus")
+            res.send(404);
+        }
     })
 })
